--- a/scipy/stats/_boost/__init__.py
+++ b/scipy/stats/_boost/__init__.py
@@ -39,3 +39,9 @@ from scipy.stats._boost.nct_ufunc import (
     _nct_isf, _nct_mean, _nct_variance,
     _nct_skewness, _nct_kurtosis_excess,
 )
+
+from scipy.stats._boost.skewnorm_ufunc import (
+    _skewnorm_pdf, _skewnorm_cdf, _skewnorm_sf, _skewnorm_ppf,
+    _skewnorm_isf, _skewnorm_mean, _skewnorm_variance,
+    _skewnorm_skewness, _skewnorm_kurtosis_excess,
+)