@zoj613 2022-08-04
Does boost implement a `fit` method for the distributions?