@mdhaber 2022-08-04
No, and I don't think there's much we can do to improve MLE besides providing a good guess, as there are no closed-form results. Fortunately, there is an analytical solution for method of moments, which would work as a guess for MLE. I can add that.