@zoj613 2022-08-05
> @tupui I overrode the `fit` method. If the chosen fit method is method of moments, it returns solutions from formulas. If the chosen fit method is MLE (default), it uses method of moments as a guess, unless the user provided guesses, in which case these are respected.
> 
> ```python
> import numpy as np
> from scipy import stats
> 
> rng = np.random.default_rng(4609813989115202851)
> a, loc, scale = -2, 3.5, 0.5
> dist = stats.skewnorm(a, loc, scale)
> rvs = dist.rvs(size=100)
> 
> stats.skewnorm.fit(rvs)  # 22.9 ms in this PR, 64.9 ms in main
> stats.skewnorm.fit(rvs, method="mm")  # 464 µs, >60s in main (never waited for it to finish)
> ```
> 
> Because this implemented method of moments, I was able to remove `skewnorm` from the skip lists for those tests.
> 
> If you'd prefer for this stuff to be in a separate PR, that's fine, as it's different from the Boost stuff and fairly technical. LMK, I can revert.

Thanks for adding this. It will definitely be useful for me in a use case that requires fitting skewnorm parameters for a large dataset.