@mdhaber 2022-08-05
> it does not bring much penalty to do so.

It was a choice intended to make `fit` with MLE faster overall, and it does in the examples I've tried. 

If you're comfortable merging, feel free. Perhaps @mckib2 can double check the Boost stuff whenever he's available, but it was pretty formulaic and matched the other examples (except that this was the first distribution for which we didn't need all of Boost's parameters because they were just location and scale) so I'm pretty confident.