@mdhaber 2022-08-05
I did a comparison of MLE fitting before and after this PR using the code below. 

<details>

```python3
import numpy as np
from scipy import stats
from time import perf_counter

rng = np.random.default_rng(4609813989115202851)
qrng = stats.qmc.Halton(6)

times = []
for i in range(1000):
    print(i)
    sample = qrng.random(random_state=qrng)
    sample = stats.qmc.scale(sample, [-3, -3, -3, 1, -1, -1], [3, 3, 3, 5, 1, 1])[0]
    a, loc, scale, size = 10**sample[:4]
    a_sign, loc_sign = np.sign(sample[4:])
    a *= a_sign
    loc *= loc_sign
    size = int(size)

    dist = stats.skewnorm(a, loc, scale)
    rvs = dist.rvs(size=size)


    try:
        tic = perf_counter()
        a2, loc2, scale2 = stats.skewnorm.fit(rvs)
        toc = perf_counter()

        nllf2 = stats.skewnorm.nnlf((a2, loc2, scale2), rvs)
        nllf = stats.skewnorm.nnlf((a, loc, scale), rvs)
        err = abs((nllf2 - nllf)/nllf)

        times.append([toc-tic, err])
        print(toc-tic, err)
    except:
        times.append([np.nan, np.nan])
        print(f"Failure: {stats.skew(rvs)}")

filename = 'skewnorm_fit_time.npz'
np.savez(filename, times=times)
```

</details>

In short, I randomly sampled distribution parameters and RVS size from a (log-scaled) wide range, generated the RVS, and measured the fitting time and "error" (relative difference between NLLF value for original and fitted parameters) 1000 times and plotted the results. (One weakness of this approach is that the RVS are actually drawn from a skew normal distribution itself, but hopefully that is representative of data that is reasonable to fit with a `skewnorm` distribution.)

Fit time is reduced. Geometric mean of time before is 0.101s , after is 0.055.

![image](https://user-images.githubusercontent.com/6570539/183121814-2b6c42a9-7df5-4689-bbcc-c4dce9f651b5.png)

I don't see much difference in the distribution of fit "error".

![image](https://user-images.githubusercontent.com/6570539/183122354-01a3d6a7-730b-4b6f-b1dd-39e79d31914d.png)

There is one thing that needs to be fixed: whenever the skewness of the sample has magnitude greater than 1, the guess for `a` is infinite, and this causes `fi`t to fail. I realized that this would happen, but I didn't realize how frequently would occur - more than 10% of the time in the test above. So I think that when sample skewness magnitude is greater than 1 and the method is "mle", we should fall back to the default guess.