@zoj613 2022-08-08
@mdhaber  Does scipy have a pre-release or nightly build that I can install to make sure of this enhancement in my code?