@mckib2 2022-08-10
@mdhaber Boost stuff looks great -- let me know if/when you run into difficulties or where you think things could improve.  We can try to revise in the next iteration of the boost gen